requirements.txt
pyspark==4.1.0
pandas==2.1.4
scikit-learn==1.3.2
kafka-python==2.0.2
numpy==1.26.2